Refactor v2 eval op dispatch
